requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
dj-database-url==0.4.1
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
kombu==4.0.0
Pillow==3.4.2
pkg-resources==0.0.0
PyJWT==1.4.2
pytz==2016.10
redis==2.10.5
rest-condition==1.0.3
vine==1.1.3
whitenoise==3.2.2
